Change buttons of result and review pages
I made some changes to the action features of a restaurant. Earlier the result showed a Menu, Order, and Review button. I've replaced all of them by a new button called View, which opens up the review page of the restaurant. I've added the Order button there. I've not placed back the Menu button because I think it is redundant with Order.